requirements.txt
absl-py==0.15.0
aiohttp==3.8.5
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
constantly==23.10.4
contourpy==1.1.1
cryptography==41.0.5
cycler==0.11.0
daphne==4.0.0
debugpy==1.8.0
decorator==5.1.1
discord.py==2.3.2
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.3.0
dlib==19.24.2
exceptiongroup==1.1.3
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.12.4
flatbuffers==1.12
fonttools==4.42.1
frozenlist==1.4.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==22.10.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
Keras==2.4.3
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mtcnn==0.1.0
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
package-name==0.1
packaging==23.1
pandas==2.1.1
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.1
tensorflow-estimator==2.5.0
termcolor==1.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.5
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.12.1
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
